Bug#22195588
SETTING NULL TO A DYNAMIC IN-MEMORY COLUMN CORRUPTS DISK COLUMN VALUES.
When a table has dynamic columns and no var sized columns in memory,
along with few other columns in disk, setting null to all those dynamic
columns will end up corrupting the values in the disk based columns.

The var part of a tuple comprises of the variable sized columns and the
dynamic columns. When a dynamic column is set to null, it is completely
ignored during the shrink phase of the tuple and is not stored. So
when a table doesn't have any var sized columns and the dynamic columns
are set to null, no var part should be written to the tuple. And during
commit, the function inspects the header_bits flag to check if VAR_PART
is present. If there is no var part present, it expects the disk column
content to be stored from the end of the fixed part.

But during shrink phase, the length of the var part is written to the
end of fix part even when there is no var part present (The length
would be zero in this case). So during commit, since the varpart is not
present, ndbd assumes disk column content is next to the fixed part and
starts reading but since the length(0) is there, the tuple gets
corrupted.

This patch fixes this problem by removing the varpart length from the
shrunken tuple when no varpart is present and ensures that the disk
column content is written next to the end of the fix part.