Bug#25188540 FIELD_TINY::VAL_INT(): ASSERTION `!TABLE || (!TABLE->READ_SET |......)' FAILED
Since 5.0 'SHOW FIELDS FROM some_view' opens the view (parses its definition)
and creates a temporary TABLE object, to get types of of the view's columns.
In case the view's definer is incorrect, SHOW FIELDS must still succeed,
so it avoids checking for privileges. That is why there are tests on
SQLCOM_SHOW_FIELDS in sql_derived.cc and sql_base.cc, per the fixes for
bug 26817 and 31662.
In 8.0, 'SHOW FIELDS' gets its information from the DD, doesn't open
the view anymore (per WL 7167). So those if() tests become unnecessary; moreover,
they are even problematic: indeed in 8.0 'SHOW FIELDS' is rewritten
to a SELECT over a system-defined view I_S.COLUMNS which itself wraps
DD tables: this system view must be opened, and if it is materialized then
underlying columns must be marked in read_set, which the old if() bypasses,
leading to the assertion failure.

Fix: remove the if() in sql_derived.cc. Also remove other if()s as
they look unnecessary too.