BUG#25461354: FLOW-CONTROL MAY FAIL TO STOP THROTTLING WHEN THE THRESHOLDS ARE LOW
When flow control is enabled, reaching a minimum flow control quota
of 1 will make Group Replication not to stop throttling once the
cause of the throttling is gone. This happens because the integer
result of 1*1.5 is 1, which makes the quota value a constant 1.

To solve the above issue we ensure that the next computed quota
value is always bigger than the previous one.