Bug#25573143 SELECT USING CTE NEVER USES THE QUERY CACHE
The Query Cache's code is meant to cache results of SELECT statements
so it has a test:
"if statement's text starts neither with a comment nor with the SELECT
word then don't search for it in the cache". It is a quick initial test
which may let to pass statements like
/* */ UPDATE ...
so there are other finer-grained tests later in the process to make sure
this UPDATE isn't searched for in the cache.
With CTEs, a SELECT may start with the WITH word, so the quick test wrongly
rejects such SELECT. Fixed by accepting the WITH word.
After that, it still didn't work for recursive CTEs. Indeed, when deciding
whether to store the query's result in the cache or not, the query
cache's code inspects all tables used by the statement, to see if
each table's content is "trackable in the cache", like this:
loop on tables:
a) if derived table, ignore this table (it's just a "read-only bunch of
rows from underlying tables", we care only for underlying tables and
will access them in the loop anyway)
b) if a tmp table, don't serve from the cache (query cache is supposed
to be global to all connections; other connections don't see the tmp table).
With a recursive CTE like
 WITH RECURSIVE cte AS (SELECT... UNION SELECT ... FROM cte) SELECT ...;
the second 'cte' reference is not a derived table, it's a recursive
reference. So (a) didn't catch it, so it came to (b), and the test
for "is tmp table" said "yes" (as it's materialized, it uses a tmp
table internally).
Fixed by extending the test on is_derived().

Added an interface is_recursive_reference().