Bug#24834622 - ASSERT IN UPDATING VIEW METADATA FOR SYS SCHEMA VIEW.
While updating view columns metadata, assert condition is hit
for the datetime type column with default value as bitmap
TABLE::read_set is not set.

TABLE has a bitmap read_set. A bit in this bitmap is set for a
field of a table used in the query. But this bitmap is not set
(for outer-most query block) while resolving derived tables.
A flag THD::derived_tables_processing is used to indicate the
derived table processing.

DDL operation on a table, views or stored function(sf) updates
a view columns metadata if there are any views referencing them.
The view columns metadata is updated as below,

  * List all the views using table, views or sf used in the DDL
    operation and even other views dependent on view(s) already
    included in the list.

  * For all views in the list,
      ** Prepare query expression for a view query with the new
         definition of table/view/sf.

      ** Update view columns metadata using the query expression
         prepared for a view query.

    (In error scenarios mark view as invalid and continue with
     next view in list.)

While preparing query expression for a view, if the derived
table processing fails in a situation like missing other table,
view or sf (while preparing the query expression of the derived
table) then the flag THD::derived_tables_processing is not reset
in the current code. This affects processing next view in the
list and we end-up not setting read_set flags in TABLE. Hence
assert condition to check bitmaps THD::read_set while accessing
fields (of type datatime in this scenario) fails.

Modified code to reset THD::derived_tables_processing flag in
in all error handling code path while resolving derived tables to
fix this issue.