Bug#25572504 - PID.SHUTDOWN PRESENT.SERVER WILL NOT RESTART                EROR WHEN STARTING >1 MYSQLD_SAFES
DESCRIPTION
===========
Script mysqld_safe isn't bringing the server up when the
latter is crashing due to some reason. The log file reads
as given below:

"$$$.pid.shutdown present. The server will not restart."

ANALYSIS
========
Whenever we run "/etc/init.d/mysql stop" (or any other
command which internally calls it), a "$$$.pid.shutdown"
file is created to verify a graceful exit. During normal
shutdown this file is expected to be removed as part of
cleanup.

As part of code refactoring done in patch to Bug#11751149,
we missed deleting the "$$$pid.shutdown" file in one of the
exit cases. As a result of which whenever a crash occurs,
none of the "$$$.pid" file and "$$$.pid.shutdown" file is
deleted. Now on finding the latter, the mysqld_safe script
refuses to restart the server and eventually comes to a halt
(which by its behaviour is correct!)

The "$$$pid.shutdown" file has to be deleted in this case.

FIX
===
Reverted the part of code change in Bug#11751149, so that
now the files are deleted in such a case.