Bug#25587256: NEWDD: FK: NEED TO SET PROPER REFERENCED COLUMN CASE
The problem was that the name of the columns referenced by
a foreign key was not converted to lower case when they
were stored in the data dictionary. This would have given
a change of behavior when the data dictionary eventually
is used for FK info in e.g. SHOW CREATE TABLE.
On current trunk, InnoDB's data dictionary is still used,
so this bug had no visible effects yet.

This patch fixes the problem by converting referenced column
names to lower case when they are stored in the data dictionary.