Bug#25608828: I_MAIN.VARIABLES-BUG21503595 FAILS               SPORADICALLY ON PB2-5.5 FOR LINUX-VALGRIND
Description: Sporadic failure of variables-bug21503595 test
on pb2-5.5 for linux-valgrind platform.

Fix: This is a issue related to libc and not related to
MySQL code. During dlclose few blocks of memory left
unfreed. This is a known issue in libc and needs to be
suppressed.

Fix: Added a valgrind suppression.