Bug#25643023	SERVER ABORTING WHILE INITIALIZING DATABASE WITH INIT SCRIPTS
If the datadir is deleted and the service restarted, the service
script will reinitialize the database, then try to set auth_socket
plugin for the now passwordless root account. The latter should not
happen, because a) It's a major change in an old GA, and b) The sql
simply fails, since it's run in bootstrap mode.