Bug#25642405 ERROR FROM DIH-BLOCK IS RETURNED AS 'ERROR 218, OUT OF LONGMESSAGEBUFFER' BY TC
During the initial phases of a SCANREQ, TC send DIGETNODESREQ's
to DIH in order to get dictionary information for each
fragment to scan.

If a REF is returned from DIH, TC never used the returned
errorCode. Instead a ZGET_DATAREC_ERROR
(Error 218: out of LongMessageBuffer) was always returned.

This patch read the 'errorIndicator' in theData[0], and 'iff != 0',
it reads the errorCode from theData[1], which is returned from
TC as the error.

Note: The above REF-protocol is a bit a awkward. However
the same pattern is used a couple of other places in the
TC-block. So I prefer reusing this pattern over redefining
the DIGETNODESREF protocol.