BUG#25256910: NODE RECOVERY FAILS WITH ERROR_CODE: 1062; HANDLER ERROR HA_ERR_FOUND_DUPP_KEY [Post-Fix]
Failures were detected on weekly-5.7 failing the execution of DROP TABLE t1.

An explanation for this problem is that at the end of the test we start the
server 2 applier again. The applier has a transaction to apply on t1 but we
never check if it is already applied or not.

In the end of the test we also do a drop on table t1, that in theory could be
certified and executed sooner than the remote transaction on t1.

So it was added a wait condition to confirm the presence of the data on
server2.