Bug#25583588 : `M_PREBUILT->TABLE->N_REF_COUNT > 0' AT                 HA_INNOBASE::UPDATE_THD IN HANDLER
Post push patch:

After pushing fix for bug#25583588, server does not show hidden tables
even with the I_S implementation that uses ST_SCHEMA_TABLE interface.
This affects the test case added by Bug#22285643, which uses
I_S.PARTITIONS to check if there exists a hidden table. This leads
to the timeout of the test.

Remove the test case and debug points introduced in the code
for this test.