Bug #25642319: ILLEGAL MIX OF COLLATIONS FOR TIME/VARCHAR
When doing comparisons, don't try to coerce the right argument to the left's
collation; instead, use the appropriate aggregations to try to find a superset.
At the same time, fix a field in DTCollation that would have undefined value when
using set().

Also, make sure UTF32 is counted as a superset of any collation, because it is.
(Previously, only utf8mb4 would have this status.)

Change-Id: Ib55502cf2a36cbca77332a1d0f2b300a2c558273