Bug#25188540 FIELD_TINY::VAL_INT(): ASSERTION `!TABLE || (!TABLE->READ_SET |......)' FAILED
While working on the fix for this bug, we noticed that
commands don't need mascarade as SHOW FIELDS anymore; this was very old
code (2004, 2007) which isn't necessary now that
opening tables doesn't imply materializing them:
- COM_FIELD_LIST calls mysqld_list_fields() which only calls
setup_materialized_table), doesn't actually create and fill it
- fill_schema_table_by_open(): same.
Note that this:
  DBUG_ASSERT(thd->lex->sql_command == SQLCOM_SHOW_FIELDS);
in get_schema_tmp_table_columns_record() is still correct, as
that function is only called by SHOW FIELDS, as can be seen in:
PT_show_fields_and_keys::contextualize():
  // If its a temporary table then use schema_table implementation.
  if (find_temporary_table(thd, tbl) != nullptr)
  {
...
    enum enum_schema_tables schema_table=
      (m_type == SHOW_FIELDS) ? SCH_TMP_TABLE_COLUMNS :
leading to reading this internal I_S table:
  {"TMP_TABLE_COLUMNS", tmp_table_columns_fields_info, create_schema_table,
    get_all_tables, make_tmp_table_columns_format,
    get_schema_tmp_table_columns_record, -1, -1, 1, 0}
which calls the function having the assertion test.
Here there is anyway no question of materializing, as it's
a temporary table.