Bug #23046302   COUNT(*) MUCH SLOWER ON 5.7 THAN 5.6
Test case fix for cmp_per_index.test which was using count(*)
for scanning the table using primary index. The removal of
Wl6742 triggered count(*) to use secondary index which was
causing the test to fail for some innodb compression levels.
Fix is to force the table to use primary index by doing
select * from table.