Bug#22671846: SENDING LARGE MESSAGES CAN TRIGGER NODE EXIT FROM GROUP
Problem:

When large messages are sent, the XCOM thread becomes busy for a long
time sending it. As this happens, mainly as the number of member
nodes grows, the alive task can trigger a group view change and GCS
can trigger a node exit due to the 5 second timeout period.

Solution:

For large messages, the select will mark the socket as readable many
times for a single message, since TCP sends a large message as
multiple fragments anyway. Each time this happens, the
acceptor_learner task is activated and will call read_msg/read_bytes.
Until now we have updated only when a complete messages has been
received, which may take a long time. By updating the server timestamp
inside the loop in read_bytes/buffered_read_bytes, the timestamp will
be updated often enough that nodes are not marked as being
unavailable.