BUG#22150112 REPLICATION BREAKS WITH NOT PRESERVE EVENTS
Problem and analysis:
Events defined as "ON COMPLETION NOT PRESERVE" can be automatically
dropped when they complete executing or when the scheduler is started.
The issue reported refers to when the event scheduler is started and
events that have passed their execution times are automatically dropped.
This automatic drop was already removing the event from the event queue
and from the mysql.event table. However, it was not being written to
the binlog and, consequently, a slave would not replicate these drop events.
This would cause an error on the slave if the master created another event
with the same name as one automatically dropped.

Fix:
When the scheduler is started, if any event is dropped automatically,
the drop is written to the binlog so that it can be replicated to
the slaves.