Bug#25642343: THE SERVER SHOULD HAVE A DEFAULT USER FOR THE SERVER SESION SERVICE
Plugins like X and Group Replication try to use the root user to
contact the server using the Server Session Service.
This approach has many drawbacks as the root user may not be in the
server or is expired.
This patch proposes a different approach by adding to the server a
default locked user that can be used by plugins.

This patch introduces:
  - The user creation on --initialize
  - The user creation on upgrade
  - MTR changes so the user is created on bootstrap