Bug #25196653: INCORRECT BEHAVIOR OF SP IN INDEX CONDITION                PUSHDOWN
Issue:
------
This problem occurs when:
1) A certain type of Stored Procedure / Function (SP) is
   used in a query's WHERE clause.
2) Such a condition is pushed to the storage engine as part
   of Index Condition Pushdown (ICP).

Root cause:
-----------
When a condition containing an SP is pushed down as part of
ICP, it might initiate new sql statements which are part of
SP. This is not allowed by some storage engines.

Fix:
----
A condition containing SP is currently restricted from being
part of ICP. But in certain cases Item_func_sp can be
converted to a constant item (Item_cache_*) and the same
check needs to applied in this case too.