Bug#24930129: REDUCE NUMBER OF DICTIONARY CACHE LOOKUPS
Post-push fix: Only check for metadata locks when DD objects
are released if EXTRA_DD_DEBUG is set instead of doing it
by default for debug builds. This check is expensive as
DD objects are cloned and then deleted for every check.
This cost was made more evident by the patch for Bug#24930129
which lead to timeouts for main.merge on Windows.

No changes for release builds.