Bug#25633041 : SHOW CREATE PROC/FUNCTION RESULT AFTER                LIVEUPGR IS DIFFERENT TO DUMPUPGR
In place upgrade from mysql-5.7 to mysql-8.0 stored servers
default client and connection character set for stored routines.
This changes the behavior of the stored routines.

Stored routines use the character set and collation in effect at
routine creation time. If user wants to change the database default
character set or collation, stored routines that use the database
defaults must be dropped and recreated so that they use the new defaults.

Fix:
Use the client and connection collation in effect when
stored procedure and stored function was created for inplace upgrade.

Updated main.dd_upgrade_test.test result file which was
recored mistakenly.