BUG#25444610 REPLICATION_APPLIER_STATUS_BY_WORKER USES COORD THREAD_ID AS INDEX WITH MTS
Problem:
When MTS is enabled, the P_S indexing functionality introduced by WL#6616
is relying on the SQL applier coordinator thread ID instead of the SQL
applier worker thread IDs.

Analysis:
The implementation of the P_S indexing functionality is only taking into
account the thread ID of the SQL thread. This is correct for STS,
because the SQL thread is the worker thread. However, for MTS, this
thread will correspond to the coordinator, while there are actual worker
threads with distinct thread IDs from the MTS coordinator.

Fix:
When MTS is enabled, use the thread ID of the correspondent worker as
index.