WL#10449 Add disk free functionality for homedir on local/remote host.
Accidental problem found with MCC when while opening console on
remote host an error is thrown (i.e. for me it was non-existent path in
.bashrc RE left there). Now code skips the error output in remote
console, if any, by taking entire output and looking for last line and
not 2nd. As a part of this, the check whether we're on Windows is
changed. Previously, we were issuing '#' to console knowing bash will
not complain while Windows CMD will. Now, we issue 'uname' (with same
failure logic) being that we need output of uname, when on *nix, anyway.

Dojo widgets constraints enlarged:
For Max. number of CPUs now 5000 (was 100)
For Max. RAM now 90 TB (was 1 TB).
This is no big deal but looks better without red exclamation marks.