Bug#25713194 MYSQL 8.0.1 - SKIP-GRANT-TABLES IS ENABLED BY DEFAULT DURING LIVE UPGRADE
During upgrade the server went into a maintenance mode which
ignored all privileges because one of the required privilege
tables was missing. With this patch the missing table is treated as
empty. A subset of the privileges will still be enforced and
default behavior will be to deny access rather than grant it.