Bug #25702850 TOO LARGE RESULT SIZE REPORTED FROM TUP-BLOCK TO LQH
The pseudo_read instruction FLUSH_AI attrInfo is used to
produce two TRANSID_AI result rows from the same row: One
going to the SPJ block with keys for fuether lookups, and the
other to the API containing the result row. The total length
of both of these rows were reported in the TUPKEYCONF signal
from TUP -> LQH. This could cause LQH to prematurely terminate
scan batches as it believed that the total number of
ScanBatch 'bytes' was about to be exhausted.

This patch change this such that only the result row going to the
API is reported as part of the 'read_length' being reported.