Bug #25189192	ERRORS WHEN RESTARTING MYSQL AFTER RENAME TABLE.
PROBLEM

While renaming table innodb doesn't update the InnoDB Dictionary table
INNODB_SYS_DATAFILES incase there is change in database while doing
rename table. Hence on a restart the server log shows error that it
couldnt find table with old path before rename which has actually been
renamed. So the errors would only vanish if we update the system
tablespace

FIX

Update the innodb dictionary table with new path in the case there is
not a change in the table but the database holding the table as well.

Reviewed-by: Jimmy Yang<Jimmy.Yang@oracle.com>
RB: 15751