WL#8963: Support for partial update of JSON in the optimizer WL#9192: Add JSON_STORAGE_SIZE / JSON_STORAGE_FREE functions
This commit introduces infrastructure in the optimizer to help the
storage engines perform partial update of JSON values more
efficiently. Partial update can be performed for UPDATE statements
such as

  UPDATE t SET json_col = JSON_SET(json_col, '$.path', 'abc')

or

  UPDATE t SET json_col = JSON_REPLACE(json_col, '$.path', 'xyz')

if the new value can be added to the JSON document by overwriting the
old value at the modified path. This is possible if the new value does
not require more storage space than the value that is replaced, or if
there is sufficiently unused spaced around the replaced value to add
the new value without increasing the size of the JSON document.

When partial update is possible, a list of differences between the
original document and the updated document is made available to the
storage engine, so that it can write only the few bytes that actually
changed, instead of rewriting the entire document, thereby reducing
the amount of I/O.

No storage engine takes advantage of this information for now.