Bug #25732787: INSTALLED HEADER FILES INCLUDE NON-INSTALLED HEADER FILES
Stop installing header files beyond those that are required for <mysql.h>
to compile. <mysql.h> is the only entry point clients should use, and it
should not include "my_config.h", shich pulls in a large amount of HAVE_*
macros into the client's namespace.

In the process, remove the my_init() call from the list of exported calls;
it's not needed, as it's implicitly called by all the other init functions,
and it's part of my_sys, which is internal-only.

Also, remove F_TO_EOF, since it required an #ifdef and was unneeded.

Change-Id: I4ce25e27e1e4e97995fe68eb220d1ade4db26301